feat: add commands.js
